Move MediaPlayerHelper to a different package because it is not UI staff
